+++ NEWS.md
@@ -5,10 +5,11 @@ New Features
 ------------
 
 * Support for correlated variables has been added
-  ([#3](https://github.com/giordano/Measurements.jl/issues/3))
+  ([#3](https://github.com/giordano/Measurements.jl/issues/3)).
 * Macro `@uncertain` enables propagation of uncertainty in any real function of
   one real argument, including those making use of C/Fortran calls
   ([#4](https://github.com/giordano/Measurements.jl/issues/4)).
+* New method supported: `typemax`.
 
 v0.0.2 (2016-06-13)
 ===================