+++ NEWS.md
@@ -8,8 +8,8 @@ New Features
   ([#3](https://github.com/giordano/Measurements.jl/issues/3)).
 * Full support for complex measurements.
 * Macro `@uncertain` enables propagation of uncertainty in any real function of
-  one real argument, including those making use of C/Fortran calls
-  ([#4](https://github.com/giordano/Measurements.jl/issues/4)).
+  one or two real arguments, including those functions making use of C/Fortran
+  calls ([#4](https://github.com/giordano/Measurements.jl/issues/4)).
 * Precompilation enabled.
 * New method supported: `typemax`.